refactor: move page location
